refactor: extract ShareModal, add share to tracking page, drop public visibility
- Extract ShareModal + copyToClipboard into shared component
  (frontend/src/components/projects/ShareModal.tsx)
- Add share icon button and 'Shared' chip badge to ProjectDetailPage header
- Remove 'Public (listed)' visibility option — no listing feature exists;
  backend now rejects anything other than 'link', modal simplified to
  single create-link flow with expiry picker
- Confirmed revoke makes slugs immediately inaccessible (404)
